refactor: enhanced error handling by adding ErrorPage, updating ErrorBoundary, and wrapping routes
