+++ src/components/ErrorComponents/index.ts
@@ -1,0 +1,4 @@
+import ErrorBoundary from './ErrorBoundary';
+import ErrorPage from './ErrorPage';
+
+export { ErrorBoundary, ErrorPage };
