+++ src/main.tsx
@@ -3,14 +3,11 @@ import { createRoot } from 'react-dom/client';
 
 import './index.css';
 import App from './App';
-import ErrorBoundary from './components/ErrorBoundary';
 
 const rootElement = document.getElementById('root');
 if (!rootElement) throw new Error('Root element not found');
 createRoot(rootElement).render(
   <StrictMode>
-    <ErrorBoundary>
-      <App />
-    </ErrorBoundary>
+    <App />
   </StrictMode>
 );
