+++ src/router/AppRouter.tsx
@@ -1,5 +1,7 @@
-import { createBrowserRouter } from 'react-router-dom';
+import { createBrowserRouter, Navigate } from 'react-router-dom';
 
+import { ErrorBoundary, ErrorPage } from '@components/ErrorComponents/';
+import { PokemonDetail } from '@components/PokemonCard';
 import AboutPage from '@pages/AboutPage';
 import { MainLayout, MainPage } from '@pages/MainPage';
 import NotFoundPage from '@pages/NotFoundPage';
@@ -8,16 +10,32 @@ export const router = createBrowserRouter([
   {
     path: '/',
     element: <MainLayout />,
+    errorElement: <ErrorPage />,
     children: [
       {
         index: true,
-        element: <MainPage />,
+        element: <Navigate to="main" replace />,
+      },
+      {
+        path: 'main',
+        element: (
+          <ErrorBoundary>
+            <MainPage />
+          </ErrorBoundary>
+        ),
+        children: [
+          {
+            path: '',
+            element: (
+              <ErrorBoundary>
+                <PokemonDetail />
+              </ErrorBoundary>
+            ),
+          },
+        ],
       },
       { path: 'about', element: <AboutPage /> },
-      {
-        path: '*',
-        element: <NotFoundPage />,
-      },
+      { path: '*', element: <NotFoundPage /> },
     ],
   },
 ]);
